Revampirize, I mean, revamp the BAT script.
Define default stages: build, publish, test, discriminator.
Those stages can be overriden by the configuration file in the
variable BISECTION_STAGES.

Each stage may or may not exist in the configuration file.

Additionally, the BAT script can run each specific stage by
specifying it in the arguments. E.g.:
  bat.sh --build db410c-b-section.conf
or:
  bat.sh --discriminator db410c-c-section.conf
or:
  bat.sh --custom_stage db410c-vv-section.conf
where the stages are defined in the configuration like:
  [build]
or:
  [discriminator]
or:
  [custom_stage]
respectively (taking the whole line in the configuration).

When the configuration is passed as the sole argument to the
BAT script, then the script will start the Git bisection and
will run all stages.

Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>